LibreOffice Calc: Refactored column merging and grouping into CalcColumnGroupBuilder.
