+++ src/main/java11/org/codehaus/mojo/license/extended/spreadsheet/CalcFileWriter.java
@@ -8,10 +8,7 @@ import java.io.File;
 import java.io.IOException;
 import java.io.OutputStream;
 import java.nio.file.Files;
-import java.util.ArrayDeque;
 import java.util.ArrayList;
-import java.util.Comparator;
-import java.util.Deque;
 import java.util.EnumMap;
 import java.util.HashMap;
 import java.util.List;
@@ -31,11 +28,9 @@ import org.codehaus.mojo.license.download.ProjectLicense;
 import org.codehaus.mojo.license.download.ProjectLicenseInfo;
 import org.codehaus.mojo.license.extended.ExtendedInfo;
 import org.codehaus.mojo.license.extended.InfoFile;
-import org.jspecify.annotations.NonNull;
 import org.odftoolkit.odfdom.doc.OdfSpreadsheetDocument;
 import org.odftoolkit.odfdom.doc.table.OdfTable;
 import org.odftoolkit.odfdom.doc.table.OdfTableCell;
-import org.odftoolkit.odfdom.doc.table.OdfTableCellRange;
 import org.odftoolkit.odfdom.doc.table.OdfTableColumn;
 import org.odftoolkit.odfdom.doc.table.OdfTableRow;
 import org.odftoolkit.odfdom.dom.OdfContentDom;
@@ -45,9 +40,6 @@ import org.odftoolkit.odfdom.dom.element.config.ConfigConfigItemMapEntryElement;
 import org.odftoolkit.odfdom.dom.element.style.StyleParagraphPropertiesElement;
 import org.odftoolkit.odfdom.dom.element.style.StyleTableCellPropertiesElement;
 import org.odftoolkit.odfdom.dom.element.style.StyleTextPropertiesElement;
-import org.odftoolkit.odfdom.dom.element.table.TableTableColumnElement;
-import org.odftoolkit.odfdom.dom.element.table.TableTableColumnGroupElement;
-import org.odftoolkit.odfdom.dom.element.table.TableTableElement;
 import org.odftoolkit.odfdom.dom.element.text.TextAElement;
 import org.odftoolkit.odfdom.dom.style.OdfStyleFamily;
 import org.odftoolkit.odfdom.dom.style.props.OdfTableColumnProperties;
@@ -102,8 +94,6 @@ import static org.codehaus.mojo.license.extended.spreadsheet.SpreadsheetUtil.get
  */
 public class CalcFileWriter {
     private static final Logger LOG = LoggerFactory.getLogger(CalcFileWriter.class);
-    /** Custom attribute name to store pending column groups, before they are added to the table. */
-    private static final String PENDING_COLUMN_GROUPS_KEY = CalcFileWriter.class.getName() + ".pendingColumnGroups";
 
     private static final String HEADER_CELL_STYLE = "headerCellStyle";
     private static final String HYPERLINK_NORMAL_STYLE = "hyperlinkNormalStyle";
@@ -170,7 +160,7 @@ public class CalcFileWriter {
                     table,
                     convertToOdfColor(SpreadsheetUtil.ALTERNATING_ROWS_COLOR),
                     formatting);
-            applyPendingColumnGroups(table);
+            CalcColumnGroupBuilder.applyPendingColumnGroups(table);
 
             try (OutputStream fileOut = Files.newOutputStream(licensesCalcOutputFile.toPath())) {
                 spreadsheet.save(fileOut);
@@ -213,12 +203,12 @@ public class CalcFileWriter {
         OdfTableRow thirdHeaderRow = table.appendRow();
 
         // Create Maven header cell
-        createMergedCellsInRow(
+        CalcColumnGroupBuilder.createMergedCellsInRow(
                 table, MAVEN_START_COLUMN, MAVEN_END_COLUMN, mavenJarRow, "Maven information", 0, HEADER_CELL_STYLE);
 
         if (hasExtendedInfo) {
             // Create JAR header cell
-            createMergedCellsInRow(
+            CalcColumnGroupBuilder.createMergedCellsInRow(
                     table,
                     EXTENDED_INFO_START_COLUMN,
                     EXTENDED_INFO_END_COLUMN,
@@ -229,11 +219,11 @@ public class CalcFileWriter {
         }
 
         // Create Maven "General" header
-        createMergedCellsInRow(
+        CalcColumnGroupBuilder.createMergedCellsInRow(
                 table, GENERAL_START_COLUMN, GENERAL_END_COLUMN, secondHeaderRow, "General", 1, HEADER_CELL_STYLE);
 
         // Create Maven "Plugin ID" header
-        createMergedCellsInRow(
+        CalcColumnGroupBuilder.createMergedCellsInRow(
                 table,
                 PLUGIN_ID_START_COLUMN,
                 PLUGIN_ID_END_COLUMN,
@@ -246,14 +236,14 @@ public class CalcFileWriter {
         setColumnWidth(table, GENERAL_END_COLUMN, GAP_WIDTH);
 
         // Create Maven "Licenses" header
-        createMergedCellsInRow(
+        CalcColumnGroupBuilder.createMergedCellsInRow(
                 table, LICENSES_START_COLUMN, LICENSES_END_COLUMN, secondHeaderRow, "Licenses", 1, HEADER_CELL_STYLE);
 
         // Gap "Plugin ID" <-> "Licenses".
         setColumnWidth(table, PLUGIN_ID_END_COLUMN, GAP_WIDTH);
 
         // Create Maven "Developers" header
-        createMergedCellsInRow(
+        CalcColumnGroupBuilder.createMergedCellsInRow(
                 table,
                 DEVELOPERS_START_COLUMN,
                 DEVELOPERS_END_COLUMN,
@@ -266,14 +256,14 @@ public class CalcFileWriter {
         setColumnWidth(table, LICENSES_END_COLUMN, GAP_WIDTH);
 
         // Create Maven "Miscellaneous" header
-        createMergedCellsInRow(
+        CalcColumnGroupBuilder.createMergedCellsInRow(
                 table, MISC_START_COLUMN, MISC_END_COLUMN, secondHeaderRow, "Miscellaneous", 1, HEADER_CELL_STYLE);
 
         // Gap "Developers" <-> "Miscellaneous".
         setColumnWidth(table, DEVELOPERS_END_COLUMN, GAP_WIDTH);
 
         if (hasExtendedInfo) {
-            createMergedCellsInRow(
+            CalcColumnGroupBuilder.createMergedCellsInRow(
                     table,
                     MANIFEST_START_COLUMN,
                     MANIFEST_END_COLUMN,
@@ -285,7 +275,7 @@ public class CalcFileWriter {
             // Gap "Miscellaneous" <-> "MANIFEST.MF".
             setColumnWidth(table, DEVELOPERS_END_COLUMN, GAP_WIDTH);
 
-            createMergedCellsInRow(
+            CalcColumnGroupBuilder.createMergedCellsInRow(
                     table,
                     INFO_NOTICES_START_COLUMN,
                     INFO_NOTICES_END_COLUMN,
@@ -297,7 +287,7 @@ public class CalcFileWriter {
             // Gap "MANIFEST.MF" <-> "Notice text files".
             setColumnWidth(table, MANIFEST_END_COLUMN, GAP_WIDTH);
 
-            createMergedCellsInRow(
+            CalcColumnGroupBuilder.createMergedCellsInRow(
                     table,
                     INFO_LICENSES_START_COLUMN,
                     INFO_LICENSES_END_COLUMN,
@@ -309,7 +299,7 @@ public class CalcFileWriter {
             // Gap "Notice text files" <-> "License text files".
             setColumnWidth(table, INFO_NOTICES_END_COLUMN, GAP_WIDTH);
 
-            createMergedCellsInRow(
+            CalcColumnGroupBuilder.createMergedCellsInRow(
                     table,
                     INFO_SPDX_START_COLUMN,
                     INFO_SPDX_END_COLUMN,
@@ -1101,167 +1091,6 @@ public class CalcFileWriter {
             cell.setValueType(VALUE_TYPE_STRING);
             cell.getOdfElement().setStyleName(styleName);
             cell.setStringValue(names[i]);
-        }
-    }
-
-    private static void createMergedCellsInRow(
-            OdfTable table,
-            int startColumn,
-            int endColumn,
-            OdfTableRow row,
-            String cellValue,
-            int rowIndex,
-            String styleName) {
-        OdfTableCell cell = createCellsInRow(startColumn, endColumn, row);
-        if (cell == null) {
-            return;
-        }
-        final boolean merge = endColumn - 1 > startColumn;
-
-        if (merge) {
-            OdfTableCellRange cellRange = table.getCellRangeByPosition(startColumn, rowIndex, endColumn - 1, rowIndex);
-            cellRange.merge();
-            addColumnGroup(table, startColumn, endColumn);
-        }
-
-        // Set value and style only after merge
-        cell.setStringValue(cellValue);
-        cell.getOdfElement().setStyleName(styleName);
-    }
-
-    private static void addColumnGroup(OdfTable table, int startColumn, int endColumn) {
-        if (endColumn - startColumn < 2) {
-            return;
-        }
-        getPendingColumnGroups(table).add(new ColumnGroup(startColumn, endColumn));
-    }
-
-    @SuppressWarnings("unchecked")
-    private static List<ColumnGroup> getPendingColumnGroups(OdfTable table) {
-        TableTableElement tableElement = table.getOdfElement();
-        List<ColumnGroup> groups = (List<ColumnGroup>) tableElement.getUserData(PENDING_COLUMN_GROUPS_KEY);
-        if (groups == null) {
-            groups = new ArrayList<>();
-            tableElement.setUserData(PENDING_COLUMN_GROUPS_KEY, groups, null);
-        }
-        return groups;
-    }
-
-    private static void applyPendingColumnGroups(OdfTable table) {
-        TableTableElement tableElement = table.getOdfElement();
-        List<ColumnGroup> groups = getPendingColumnGroups(table);
-        if (groups.isEmpty()) {
-            return;
-        }
-
-        int columnCount = table.getColumnCount();
-        for (int i = 0; i < columnCount; i++) {
-            // Auto-Extend the number of columns, so no columns are missing.
-            table.getColumnByIndex(i);
-        }
-
-        List<TableTableColumnElement> columnElements = getDirectColumnElements(tableElement);
-        if (columnElements.isEmpty()) {
-            tableElement.setUserData(PENDING_COLUMN_GROUPS_KEY, null, null);
-            return;
-        }
-
-        ColumnGroup rootGroup = buildColumnGroupTree(columnElements.size(), groups);
-        Node firstNonColumnNode = tableElement.getFirstChild();
-        while (firstNonColumnNode instanceof TableTableColumnElement) {
-            firstNonColumnNode = firstNonColumnNode.getNextSibling();
-        }
-
-        for (TableTableColumnElement columnElement : columnElements) {
-            tableElement.removeChild(columnElement);
-        }
-
-        Node rootFragment = tableElement.getOwnerDocument().createDocumentFragment();
-        appendColumns(rootFragment, tableElement, columnElements, rootGroup);
-        if (firstNonColumnNode == null) {
-            tableElement.appendChild(rootFragment);
-        } else {
-            tableElement.insertBefore(rootFragment, firstNonColumnNode);
-        }
-        tableElement.setUserData(PENDING_COLUMN_GROUPS_KEY, null, null);
-    }
-
-    private static List<TableTableColumnElement> getDirectColumnElements(TableTableElement tableElement) {
-        List<TableTableColumnElement> columnElements = new ArrayList<>();
-        for (Node child = tableElement.getFirstChild(); child != null; child = child.getNextSibling()) {
-            if (child instanceof TableTableColumnElement) {
-                columnElements.add((TableTableColumnElement) child);
-            }
-        }
-        return columnElements;
-    }
-
-    private static ColumnGroup buildColumnGroupTree(int columnCount, List<ColumnGroup> groups) {
-        List<ColumnGroup> sortedGroups = new ArrayList<>(groups);
-        sortedGroups.sort(Comparator.comparingInt((ColumnGroup columnGroup) -> columnGroup.startColumn));
-
-        ColumnGroup rootGroup = new ColumnGroup(0, columnCount);
-        Deque<ColumnGroup> stack = new ArrayDeque<>();
-        stack.push(rootGroup);
-        for (ColumnGroup group : sortedGroups) {
-            ColumnGroup parentGroup = stack.peek();
-            while (parentGroup != null && group.startColumn >= parentGroup.endColumn) {
-                stack.pop();
-                parentGroup = stack.peek();
-            }
-            if (parentGroup == null
-                    || group.startColumn < parentGroup.startColumn
-                    || group.endColumn > parentGroup.endColumn) {
-                throw new IllegalArgumentException("Column groups must be properly nested without crossings.");
-            }
-            parentGroup.nestedGroups.add(group);
-            stack.push(group);
-        }
-        return rootGroup;
-    }
-
-    private static void appendColumns(
-            Node parentNode,
-            TableTableElement tableElement,
-            List<TableTableColumnElement> columnElements,
-            @NonNull ColumnGroup group) {
-        int columnIndex = group.startColumn;
-        for (ColumnGroup nestedGroup : group.nestedGroups) {
-            while (columnIndex < nestedGroup.startColumn) {
-                parentNode.appendChild(columnElements.get(columnIndex++));
-            }
-            TableTableColumnGroupElement groupElement = tableElement.newTableTableColumnGroupElement();
-            appendColumns(groupElement, tableElement, columnElements, nestedGroup);
-            parentNode.appendChild(groupElement);
-            columnIndex = nestedGroup.endColumn;
-        }
-        while (columnIndex < group.endColumn) {
-            parentNode.appendChild(columnElements.get(columnIndex++));
-        }
-    }
-
-    private static OdfTableCell createCellsInRow(int startColumn, int exclusiveEndColumn, OdfTableRow inRow) {
-        OdfTableCell firstCell = null;
-        for (int i = startColumn; i < exclusiveEndColumn; i++) {
-            OdfTableCell cell = inRow.getCellByIndex(i);
-            if (i == startColumn) {
-                firstCell = cell;
-            }
-        }
-        return firstCell;
-    }
-
-    private static final class ColumnGroup {
-        /** Inclusive start column index. */
-        private final int startColumn;
-        /** Exclusive end column index, i.e. the first column index that is not part of this group. */
-        private final int endColumn;
-
-        private final List<ColumnGroup> nestedGroups = new ArrayList<>();
-
-        private ColumnGroup(int startColumn, int endColumn) {
-            this.startColumn = startColumn;
-            this.endColumn = endColumn;
         }
     }
